refactor: move donations route to donation history tab domain
