+++ src/DonorProfiles/ServiceProvider.php
@@ -7,10 +7,9 @@ use Give\Helpers\Hooks;
 use Give\DonorProfiles\Shortcode as Shortcode;
 use Give\DonorProfiles\Block as Block;
 use Give\DonorProfiles\App as App;
-use Give\DonorProfiles\Routes\DonationsRoute;
-use Give\DonorProfiles\Routes\ProfileRoute;
 
 use Give\DonorProfiles\Tabs\ProfileTab\Tab as ProfileTab;
+use Give\DonorProfiles\Tabs\DonationHistoryTab\Tab as DonationHistoryTab;
 
 use Give\DonorProfiles\Tabs\TabsRegister;
 
@@ -26,8 +25,6 @@ class ServiceProvider implements ServiceProviderInterface {
 		give()->singleton( App::class );
 		give()->singleton( Shortcode::class );
 
-		give()->bind( DonationsRoute::class );
-
 		if ( function_exists( 'register_block_type' ) ) {
 			give()->singleton( Block::class );
 		}
@@ -39,12 +36,14 @@ class ServiceProvider implements ServiceProviderInterface {
 	public function boot() {
 		Hooks::addAction( 'init', Shortcode::class, 'addShortcode' );
 
+		// Register Tabs
 		Hooks::addAction( 'init', ProfileTab::class, 'registerTab' );
+		Hooks::addAction( 'init', DonationHistoryTab::class, 'registerTab' );
 
-		Hooks::addAction( 'wp_enqueue_scripts', Shortcode::class, 'loadFrontendAssets' );
 		Hooks::addAction( 'wp_enqueue_scripts', TabsRegister::class, 'enqueueTabAssets' );
 		Hooks::addAction( 'rest_api_init', TabsRegister::class, 'registerTabRoutes' );
-		Hooks::addAction( 'rest_api_init', DonationsRoute::class, 'registerRoute' );
+
+		Hooks::addAction( 'wp_enqueue_scripts', Shortcode::class, 'loadFrontendAssets' );
 
 		if ( function_exists( 'register_block_type' ) ) {
 			Hooks::addAction( 'init', Block::class, 'addBlock' );
