+++ src/DonorProfiles/resources/js/app/tabs/donation-history/utils/index.js
@@ -7,11 +7,12 @@ export const fetchDonationsDataFromAPI = () => {
 	const { dispatch } = store;
 
 	dispatch( setQuerying( true ) );
-	axios.get( getAPIRoot() + 'give-api/v2/donor-profile/donations', {
-		headers: {
-			'X-WP-Nonce': getAPINonce(),
-		},
-	} )
+	axios.post( getAPIRoot() + 'give-api/v2/donor-profile/donations', {},
+		{
+			headers: {
+				'X-WP-Nonce': getAPINonce(),
+			},
+		} )
 		.then( ( response ) => response.data )
 		.then( ( data ) => {
 			dispatch( setDonations( data.donations ) );
